added first version of 'depot' and example notebook
A depot can store a record of the associated accounts and loads the accounts
when (re-)created. Loading means that the accounts are created as if they would
be created on the command line. This ensures that their creation mechanisms,
including the automatic loading of new transactions, works. Otherwise the depot
has currently no actions.

I use the example notebook for testing.