comprehensive changes to make conversions fail gracefully
The main idea is to let the 'convert'-function return nan whenever it cannot
convert between the desired symbols for whatever reason. This is to allow depot
overviews to be calculated without errors.

I catch all errors during conversion with external packages and pass them on as
warnings. Warnings will also be given when symbols are used for which no
conversion is implemented, yet.