find periodically reoccurring transactions
I added functionality to identify transactions that reoccur every month, or
every year: accounts now have a corresponding function find_reoccurring(). To
implement this I re-structured code for get_month and get_year and introduced a
new flexible_transaction_matcher. The latter allows partial matches of selected
transaction fields to count as match so that some variability in reoccurring
transactions is handled. As the implementation makes heavy use of loops, it's
performance is not ideal: months are analysed quickly, but years take a while.