more consistent handling of missing values in str-columns
When e.g. a description is missing in a transaction, this is encoded
as NaN. Previous code returned all transactions with missing values
in a search column. This was unintended. I now handle NaN consistently
for all columns in are_equal_or_nan. Consequently, missing values are
now not matched in a search anymore.