fix: inexact boundaries for months
I only replaced the day from timestamp.now to 1 for the beginning of a
month, but if you run this in the evening, transactions from the
beginning of the first of the month will be missed while other
transactions from the beginning of the next month will be added.

The code now is a bit clearer using pandas and ensures that the
exact month boundaries are used.